Reorganize scripts related to Presenter layer
