+++ Assets/Scripts/Presenter/Figure.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba13869ad73e9630584dcc4b37a6cafe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
