+++ Assets/Scripts/Presenter/Figure/CirclePresenter.cs
@@ -4,7 +4,7 @@ using StereoApp.Model;
 using StereoApp.Presenter.Base;
 using UnityEngine;
 
-namespace StereoApp.Presenter
+namespace StereoApp.Presenter.Figure
 {
     public class CirclePresenter : FigurePresenter<Circle>
     {
