+++ Assets/Scripts/Presenter/Figure/ConicalFrustumPresenter.cs
@@ -1,10 +1,9 @@
-using System;
 using System.ComponentModel;
 using StereoApp.Model.Interfaces;
 using StereoApp.Presenter.Base;
 using UnityEngine;
 
-namespace StereoApp.Presenter
+namespace StereoApp.Presenter.Figure
 {
     public class ConicalFrustumPresenter : GeneratedMeshPresenter<IConicalFrustum>
     {
