+++ Assets/Scripts/Presenter/Figure/PointExtensions.cs.meta
