+++ Assets/Scripts/Presenter/Figure/SpherePresenter.cs
@@ -3,7 +3,7 @@ using StereoApp.Model;
 using StereoApp.Presenter.Base;
 using UnityEngine;
 
-namespace StereoApp.Presenter
+namespace StereoApp.Presenter.Figure
 {
     public class SpherePresenter : GeneratedMeshPresenter<Sphere>
     {
