+++ Assets/Scripts/Presenter/UIHandlers/ToolbarMenu/Coordinates.cs.meta
